backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
razorpay
openai
celery
redis
python-dotenv
pydantic-settings
alembic
yt-dlp
groq
google-auth
email-validator
boto3
